refactor: move validation to its own layer
