+++ src/infra/adapters/phone-validator-adapter.ts
@@ -1,4 +1,4 @@
-import { PhoneValidator } from '../../presentation/protocols/phone-validator'
+import { PhoneValidator } from '../../validation/protocols/phone-validator'
 import PhoneBrValidator from 'br-phone-validator-lib'
 export class PhoneValidatorAdapter implements PhoneValidator {
   isValid (phone: string): boolean {
