+++ src/main/factories/controllers/signup/signup-validation-factory.spec.ts
@@ -1,10 +1,10 @@
 import { makeSignUpValidation } from './signup-validation-factory'
-import { ValidationComposite, CompareFieldValidation, RequiredFieldValidation, EmailValidation, PhoneValidation } from '../../../../presentation/helpers/validators'
+import { ValidationComposite, CompareFieldValidation, RequiredFieldValidation, EmailValidation, PhoneValidation } from '../../../../validation/validators'
 import { Validation } from '../../../../presentation/protocols/validation'
-import { EmailValidator } from '../../../../presentation/protocols/email-validator'
-import { PhoneValidator } from '../../../../presentation/protocols/phone-validator'
+import { EmailValidator } from '../../../../validation/protocols/email-validator'
+import { PhoneValidator } from '../../../../validation/protocols/phone-validator'
 
-jest.mock('../../../../presentation/helpers/validators/validation-composite')
+jest.mock('../../../../validation/validators/validation-composite')
 
 const makePhoneValidator = (): PhoneValidator => {
   class PhoneValidatorStub implements PhoneValidator {
