+++ src/validation/validators/compare-field-validation.spec.ts
@@ -1,5 +1,5 @@
 import { CompareFieldValidation } from './compare-field-validation'
-import { InvalidParamError } from '../../errors'
+import { InvalidParamError } from '../../presentation/errors'
 
 const makeSut = (): CompareFieldValidation => {
   return new CompareFieldValidation('field', 'fieldToCompare')
