+++ src/validation/validators/compare-field-validation.ts
@@ -1,5 +1,5 @@
-import { Validation } from '../../protocols/validation'
-import { InvalidParamError } from '../../errors'
+import { Validation } from '../../presentation/protocols/validation'
+import { InvalidParamError } from '../../presentation/errors'
 
 export class CompareFieldValidation implements Validation {
   private readonly fieldName: string
