+++ src/validation/validators/email-validation.ts
@@ -1,6 +1,6 @@
-import { Validation } from '../../protocols/validation'
-import { InvalidParamError } from '../../errors'
-import { EmailValidator } from '../../protocols/email-validator'
+import { Validation } from '../../presentation/protocols/validation'
+import { InvalidParamError } from '../../presentation/errors'
+import { EmailValidator } from '../protocols/email-validator'
 
 export class EmailValidation implements Validation {
   private readonly fieldName: string
