+++ src/validation/validators/required-field-validation.ts
@@ -1,5 +1,5 @@
-import { Validation } from '../../protocols/validation'
-import { MissingParamError } from '../../errors'
+import { Validation } from '../../presentation/protocols/validation'
+import { MissingParamError } from '../../presentation/errors'
 
 export class RequiredFieldValidation implements Validation {
   private readonly fieldName: string
