+++ src/validation/validators/validation-composite.spec.ts
@@ -1,6 +1,6 @@
 import { ValidationComposite } from './validation-composite'
-import { MissingParamError } from '../../errors'
-import { Validation } from '../../protocols/validation'
+import { MissingParamError } from '../../presentation/errors'
+import { Validation } from '../../presentation/protocols/validation'
 
 const makeValidation = (): Validation => {
   class ValidationStub implements Validation {
